app/requirements.txt
flasgger
mysql